pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "datetime>=5.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "sendgrid>=6.12.4",
    "streamlit>=1.46.1",
    "uuid>=1.30",
    "validators>=0.35.0",
]
